[@dhealthdapps/backend] test(oauth): create & fix tests for oauth scope
